+++ src/explore/explorer.rs
@@ -30,7 +30,7 @@ use super::input_box::{InputBoxState, InputBoxWidget};
 use super::limit_widget::LimitWidget;
 use super::message_box::{MessageBox, MessageBoxType};
 use super::utils::{StylePalette, Utils};
-use super::view::{SavedView, ViewsListState, ViewsListWidget};
+use super::view::{SavedView, ViewsListState, ViewsListWidget, RECENT_ALERTS, SAVED_VIEWS};
 use super::{column_frame::ColumnFrame, grid_frame::GridFrame};
 
 #[derive(Clone, Copy, Debug, PartialEq)]
@@ -74,7 +74,7 @@ pub struct Explorer {
 
 impl Explorer {
     pub fn new() -> Self {
-        Self {
+        let mut explorer = Self {
             model: DomainModel::new(),
             focus: Focus::DataGrid,
             needs_query_refresh: true,
@@ -87,7 +87,11 @@ impl Explorer {
             message_box: None,
             input_box_state: None,
             views_state: None,
-        }
+        };
+
+        explorer.apply_view(&RECENT_ALERTS);
+
+        explorer
     }
 
     fn draw(
@@ -213,14 +217,15 @@ impl Explorer {
             );
 
             let tabs_highlight = match self.focus {
-                Focus::Tabs => StylePalette::TabHighlight.style(),
-                _ => StylePalette::Tab.style(),
+                Focus::Tabs => StylePalette::TabFocusHighlight.style(),
+                _ => StylePalette::TabHighlight.style(),
             };
 
             let tabs_block = Block::default();
 
             let tabs = Tabs::new(titles)
                 .block(tabs_block)
+                .style(StylePalette::Tab.style())
                 .highlight_style(tabs_highlight)
                 .divider(symbols::DOT)
                 .select(self.model.current_type().index());